changes in server, auth, users. functioning login.

+++ server.js
@@ -11,12 +11,11 @@ const flash = require("connect-flash");
 const userInViews = require("./config/middleware/userInViews");
 const authRouter = require("./routes/auth");
 const usersRouter = require("./routes/users");
-const app = express();
 const exphbs = require("express-handlebars");
 const db = require("./models");
 const Handlebars = require("handlebars");
 const {
-  allowInsecurePrototypeAccess
+  allowInsecurePrototypeAccess,
 } = require("@handlebars/allow-prototype-access");
 const htmlRouter = require("./routes/html-routes");
 // const apiroutes = require("./routes/api-routes");
@@ -24,12 +23,14 @@ const htmlRouter = require("./routes/html-routes");
 const strategy = new Auth0Strategy(
   {
     domain: process.env.AUTH_DOMAIN,
-    clientID:process.env.CLIENT_ID,
+    clientID: process.env.CLIENT_ID,
     clientSecret: process.env.APP_SESSION_SECRET,
     callbackURL:
-      process.env.AUTH0_CALLBACK_URL || "http://localhost:8080/callback"
+      process.env.AUTH0_CALLBACK_URL || "http://localhost:8080/callback",
   },
   function (accessToken, refreshToken, extraParams, profile, done) {
+    console.log(accessToken, refreshToken, extraParams, profile, done);
+    // callback get route
     // accessToken is the token to call Auth0 API (not needed in the most cases)
     // extraParams.id_token has the JSON Web Token
     // profile has all the information from the user
@@ -50,15 +51,10 @@ passport.deserializeUser(function (user, done) {
   done(null, user);
 });
 
+const app = express();
+
 app.use(express.urlencoded({ extended: true }));
 app.use(express.json());
-
-app.use(passport.initialize());
-app.use(passport.session());
-
-app.use(express.static(join(__dirname, "public")));
-
-
 
 
 
@@ -66,7 +62,7 @@ app.engine(
   "handlebars",
   exphbs({
     defaultLayout: "main",
-    handlebars: allowInsecurePrototypeAccess(Handlebars)
+    handlebars: allowInsecurePrototypeAccess(Handlebars),
   })
 );
 app.set("view engine", "handlebars");
@@ -76,12 +72,11 @@ app.use(cookieParser());
 
 // config express-session
 var sess = {
-  secret: "CHANGE THIS SECRET",
+  secret: "random buncha shit",
   cookie: {},
   resave: false,
-  saveUninitialized: true
+  saveUninitialized: true,
 };
-
 
 if (app.get("env") === "production") {
   // If you are using a hosting provider which uses a proxy (eg. Heroku),
@@ -92,13 +87,15 @@ if (app.get("env") === "production") {
   // Ref: https://github.com/auth0/passport-auth0/issues/70#issuecomment-480771614
   // Ref: https://www.npmjs.com/package/express-session#cookiesecure
   // app.set("trust proxy", 1);
-  
+
   sess.cookie.secure = true; // serve secure cookies, requires https
 }
 
 app.use(session(sess));
 
-
+app.use(passport.initialize());
+app.use(passport.session());
+app.use(express.static(join(__dirname, "public")));
 
 app.use(flash());
 
@@ -129,22 +126,21 @@ app.use(function (req, res, next) {
 
 // Development error handler
 // Will print stacktrace
-// if (app.get("env") === "development") {
-//   // eslint-disable-next-line no-unused-vars
-//   app.use(function (err, req, res, next) {
-//     res.status(err.status || 500);
-//     res.render("error", {
-//       message: err.message,
-//       error: err
-//     });
-//   });
-// }
+if (app.get("env") === "development") {
+  // eslint-disable-next-line no-unused-vars
+  app.use(function (err, req, res, next) {
+    res.status(err.status || 500);
+    res.render("error", {
+      message: err.message,
+      error: err,
+    });
+  });
+}
 
-db.sequelize.sync({ force: true }).then(function() {
-  app.listen(PORT, function() {
+db.sequelize.sync({ force: true }).then(function () {
+  app.listen(PORT, function () {
     console.log("App listening on PORT " + PORT);
   });
 });
-  
 
 module.exports = app;
